+++ internal/server/server_custom_handlers.go
@@ -16,6 +16,7 @@ import (
 func addCustomRoutes(m *http.ServeMux, s *server) {
 	m.HandleFunc("GET /custom-api/v1/lxc/{id}/ip", getLXCIPHandlerV1(s))
 	m.HandleFunc("POST /custom-api/v1/lxc/{id}/exec", postLXCCMDHandlerV1(s))
+	m.HandleFunc("POST /custom-api/v1/lxc/{id}/exec-async", postLXCCMDAsyncHandlerV1(s))
 }
 
 // Get LXC IP godoc
@@ -106,7 +107,7 @@ func getLXCIPHandlerV1(s *server) http.HandlerFunc {
 	})
 }
 
-// Get LXC IP godoc
+// POST LXC CMD godoc
 //
 //	@Tags		LXC
 //	@Summary	Executes a command inside a proxmox lxc.
@@ -224,5 +225,145 @@ func postLXCCMDHandlerV1(s *server) http.HandlerFunc {
 		w.Write(b)
 		log.Println(r.Method, r.URL.Path, "succeeded")
 		return
+	})
+}
+
+// POST LXC CMD Async godoc
+//
+//	@Tags		LXC
+//	@Summary	Executes a command inside a proxmox lxc and returns an id used to query the result.
+//	@Description	Requires VM.Console scope. Run a command inside a proxmox lxc. It stores the exit code and it's output in a temprary file in the /tmp dir. Once that file is read, it is removed.
+//	@Param		id	path	int	true	"vmid"
+//	@Param		request body	apidef.PostLXCExecRequest true "command input"
+//	@Accept		json
+//	@Produce	json
+//	@Success	200	{object}	apidef.PostLXCExecResponse
+//	@Failure	400	{object}	errors.HTTPError
+//	@Failure	401	{object}	errors.HTTPError
+//	@Failure	404	{object}	errors.HTTPError
+//	@Failure	500	{object}	errors.HTTPError
+//	@Router		/lxc/{id}/exec [post]
+func postLXCCMDAsyncHandlerV1(s *server) http.HandlerFunc {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		log.Println(r.Method, r.URL.Path, "started")
+		authorized, err := s.IsUserAuthorized(r, "VMS", "VM.Console")
+		if err != nil {
+			httperr := errors.NewHTTPError(
+				http.StatusInternalServerError,
+				"unable to authorize request",
+				err,
+			)
+			b, _ := httperr.Marshall()
+			log.Println(r.Method, r.URL.Path, "failed", string(b))
+			return
+		}
+
+		if !authorized {
+			httperr := errors.NewHTTPError(
+				http.StatusUnauthorized,
+				"unauthorized",
+				err,
+			)
+			b, _ := httperr.Marshall()
+			log.Println(r.Method, r.URL.Path, "failed", string(b))
+			return
+		}
+
+		rawId := r.PathValue("id")
+		id, err := strconv.Atoi(rawId)
+		if err != nil {
+			httperr := errors.NewHTTPError(
+				http.StatusBadRequest,
+				"id property from path is not a number",
+				err,
+			)
+			b, _ := httperr.Marshall()
+			log.Println(r.Method, r.URL.Path, "failed", string(b))
+			return
+		}
+
+		b, err := io.ReadAll(r.Body)
+		if err != nil {
+			httperr := errors.NewHTTPError(
+				http.StatusBadRequest,
+				"missing request body",
+				err,
+			)
+			b, _ := httperr.Marshall()
+			log.Println(r.Method, r.URL.Path, "failed", string(b))
+			return
+		}
+		in := apidef.PostLXCExecRequest{}
+		if err = json.Unmarshal(b, &in); err != nil {
+			httperr := errors.NewHTTPError(
+				http.StatusBadRequest,
+				"invalid request body",
+				err,
+			)
+			b, _ := httperr.Marshall()
+			log.Println(r.Method, r.URL.Path, "failed", string(b))
+			return
+		}
+
+		if in.CMD == "" {
+			httperr := errors.NewHTTPError(
+				http.StatusBadRequest,
+				"found empty or missing .cmd property",
+				nil,
+			)
+			b, _ := httperr.Marshall()
+			log.Println(r.Method, r.URL.Path, "failed", string(b))
+			return
+		}
+
+		execId, err := s.models.CMDExecution.New()
+		if err != nil {
+			httperr := errors.NewHTTPError(
+				http.StatusInternalServerError,
+				"unable to store execution id",
+				err,
+			)
+			b, _ := httperr.Marshall()
+			log.Println(r.Method, r.URL.Path, "failed", string(b))
+			httperr.WriteResponse(w)
+			return
+		}
+
+		res := apidef.PostLXCExecAsyncResponse{ID: execId}
+		b, _ = json.Marshal(res)
+		w.WriteHeader(http.StatusOK)
+		w.Header().Add("content-type", "application/json")
+		w.Write(b)
+		log.Println(r.Method, r.URL.Path, "succeeded")
+		go func() {
+			// async proccess
+			out, exitCode, err := pveutils.ExecLXCCmd(id, in.Shell, in.CMD)
+			if err != nil {
+				if err := s.models.CMDExecution.SetFailed(execId, err); err != nil {
+					log.Println(r.Method, r.URL.Path, "failed_async", err.Error())
+				}
+				return
+			}
+
+			if exitCode == -1 {
+				httperr := errors.NewHTTPError(
+					http.StatusNotFound,
+					"lxc not found",
+					nil,
+				)
+				if err := s.models.CMDExecution.SetFailed(execId, httperr); err != nil {
+					log.Println(r.Method, r.URL.Path, "failed_async", err.Error())
+				}
+				return
+			}
+
+			if err := s.models.CMDExecution.SetSucceeded(execId, out, exitCode); err != nil {
+				log.Println(r.Method, r.URL.Path, "failed_async", err.Error())
+				return
+			}
+			log.Println(r.Method, r.URL.Path, "succeeded_async")
+			return
+		}()
+
 	})
 }