+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### PVE API wrapper
+#### Added
+- `POST /custom-api/v1/lxc/{id}/exec-async` endpoint.
 
 ## [v0.6.1]
 ### PVE API client